requirements.txt
flask==1.0.2
Flask-WTF==0.14.2
Jinja2==2.10
google-cloud-storage==1.13.0
google-cloud-firestore==0.30.0
google-cloud-pubsub==0.39.0
firebase-admin==2.14.0
tensorflow==2.0.0-beta1
tensorflow_hub==0.4.0

